Revert "GDScript: Implement get_dependencies()"
This reverts commit dc73440f899e6f32de748787e946ad762771fda0.

This commit in some form is needed to fix handling of dependencies on
export, but as it's also used for import, it's exposing some pre-existing
issues which we need to solve first.

So reverting for now to give ourselves time to iron this out for a future
Godot release.

Fixes #91726.